Avatar refactor * Fix width of icons in avatar * merge in 1 file
