--- src/components/avatar/avatarConstants.js
@@ -1,13 +1,0 @@
-import {types as ICON_TYPE, colors as ICON_COLOR} from '../icons/Icon';
-
-const SIZE = {
-  XSMALL: {className: 'xsmall', iconSize: 14},
-  SMALL: {className: 'small', iconSize: 22},
-  NORMAL: {className: 'normal', iconSize: 32},
-  LARGE: {className: 'large', iconSize: 46},
-  XLARGE: {className: 'xlarge', iconSize: 62},
-  XXLARGE: {className: 'xxlarge', iconSize: 94},
-  XXXLARGE: {className: 'xxxlarge', iconSize: 118}
-};
-
-export {SIZE, ICON_TYPE, ICON_COLOR};
